YUK-11646 Dispatch RawPointHistory changeids roll over at 2 billion -Changed type of changeID from int to __int64 to hold values greater than 2^31

+++ yukon-server/COMMON/include/utility.h
@@ -25,7 +25,7 @@ IM_EX_CTIBASE LONG LMControlHistoryIdGen(bool force = false);
 IM_EX_CTIBASE LONG CommErrorHistoryIdGen(bool force = false);
 IM_EX_CTIBASE LONG VerificationSequenceGen(bool force = false, int force_value = 0);
 IM_EX_CTIBASE int  DynamicPaoStatisticsIdGen();
-IM_EX_CTIBASE INT  ChangeIdGen(bool force = false);
+IM_EX_CTIBASE __int64 ChangeIdGen(bool force = false);
 IM_EX_CTIBASE INT  SystemLogIdGen();
 IM_EX_CTIBASE INT  CCEventActionIdGen(LONG capBankPointId);
 IM_EX_CTIBASE INT  CCEventLogIdGen();
